Apply CL color range

Apply chroma (C) and luminance (L) ranges to a vector of R colors

applyCLrange(x, Crange = NULL, Lrange = NULL, Cgrey = 5,
  fixYellow = TRUE, CLmethod = c("scale", "floor", "expand"), ...)

Arguments

x
vector of R colors

Crange
NULL or numeric vector with minimum and maximum allowed values for the chroma (C) component.

Lrange
NULL or numeric vector with minimum and maximum allowed values for the luminance (L) component.

Cgrey
numeric chroma (C) value, which defines grey colors at or below this chroma. Any colors at or below the grey cutoff will have their C values unchanged. This mechanism prevents converting black to red, for example. To disable the effect, set Cgrey=-1.

fixYellow
boolean indicating whether to "fix" the darkening of yellow, which otherwise turns to green. Instead, since JAM can, JAM will make the yellow slightly more golden before darkening, which is achieved by calling fixYellowHue().

CLmethod
character string indicating how to alter values outside the respective Crange and Lrange ranges. "scale" will rescale values only if any are outside of range, and will rescale the full range of c(Crange, Cvalues) to c(Crange). In this way, only values outside the range are rescaled. "floor" will apply a fixed cutoff, any values outside the range are set to equal the range boundary itself. "expand" will rescale all values so the range is equal to Crange.

...
additional argyments are passed to fixYellowHue() when fixYellow is TRUE.

Value

vector of colors after applying the chroma (C) and luminance (L) ranges.

Details

This function is primarily intended to restrict the range of brightness values so they contrast with a background color, particularly when the background color may be bright or dark in different scenarios.

The C and L values are defined by colorspace::polarLUV(), where C is typically restricted to 0..100 and L is typically 0..100. For some colors, values above 100 are allowed.

Values are restricted to the given numeric range using one of three methods, set via the CLmethod argument.

As an example, consider what should be done when Crange <- c(10,70) and the C values are Cvalues <- c(50, 60, 70, 80).

"floor" uses jamba::noiseFloor() to apply fixed cutoffs at the minimum and maximum range. This method has the effect of making all values outside the range into an equal final value.

"scale" will apply jamba::normScale() to rescale only values outside the given range. For example, c(Crange, Cvalues) as the initial range, it constrains values to c(Crange). This method has the effect of maintaining the relative difference between values.

"expand" will simply apply jamba::normScale() to fit the values to the minimum and maximum range values. This method has the effect of forcing colors to fit the full numeric range, even when the original differences between values were small.

In case (1) above, Cvalues will become c(50, 60, 70, 70). In case (2) above, Cvalues will become c(44, 53, 61, 70) In case (3) above, Cvalues will become c(10, 30, 50, 70)

Note that colors with C (chroma) values less than Cgrey will not have the C value changed, in order to maintain colors at a greyscale, without colorizing them. Particularly for pure grey, which has C=0, but is still required to have a hue H, it is important not to increase C.

See also

Examples

applyCLrange(c("red","blue","yellow"), Lrange=c(0,60));

#> Warning: package 'colorspace' was built under R version 3.5.2

#>      col_v1      col_v2      col_v3 
#> "#7E0000FF" "#404000FF" "#704600FF"