convert date DDmmmYYYY to Date

convert date DDmmmYYYY to Date

Usage

asDate(getDateValues, dateFormat = "%d%b%Y", ...)

Arguments

getDateValues

character date, in format recognized by dateFormat

dateFormat

character string representing the recognized date format, by default "DDmmmYYYY", which recognizes "23aug2007".

...

additional parameters are ignored.

Value

Date object

Details

This function converts a text date string to Date object, mainly to allow date-related math operations, for example difftime.

See also

Examples

asDate(getDate());
#>    22mar2025 
#> "2025-03-22"