convert R color to HCL color matrix
Source: R/jamba-colors.r

convert R color to HCL color matrix

col2hcl(x, maxColorValue = 255, ...)

Arguments

x
R compatible color, either a color name, or hex value, or a mixture of the two. Any value compatible with col2rgb.

maxValue
numeric maximum value to return, useful when the downstream alpha range should be 255. By default maxValue=1 is returned.

Details

This function takes an R color and converts to an HCL matrix, using the colorspace package, and RGB and polarLUV functions. It is also used to maintain alpha transparency, to enable interconversion via other color manipulation functions as well.

See also

Examples

col2hcl("#FF000044")

#>       #FF000044
#> L        53.241
#> C       179.041
#> H        12.174
#> alpha     0.267