Compare two functions line by line using diff

Compare two functions line by line using diff

diff_functions(
  f1,
  f2,
  width = NULL,
  wide = TRUE,
  trim_whitespace = FALSE,
  verbose = FALSE,
  ...
)

Arguments

f1

function expected to be passed by function name.

f2

function expected to be passed by function name.

width

integer number of columns to allow in output.

wide

logical indicating whether to use wide output from diff, which prints the two functions side by side.

trim_whitespace

logical indicating whether to trim leading and trailing whitespace on each function line to avoid differences defined only by the indentation of text.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

None. This function is called for its output, which can be captured with utils::capture.output().

Details

This function is a utility wrapper around linux commandline tool diff, used to compare the text representation of two R functions.

See also

Examples

diff_functions(base::grepl, base::agrepl)
#> ##  (18:36:58) 31May2022:   f1_name: base..grepl 
#> ##  (18:36:58) 31May2022:   f1_file: /var/folders/9z/y1_0kf711fx3lgc68w1k27xr0000gs/T//RtmpSSZc9j/base..grepl.R 
#> ##  (18:36:58) 31May2022:   f2_name: base..agrepl 
#> ##  (18:36:58) 31May2022:   f2_file: /var/folders/9z/y1_0kf711fx3lgc68w1k27xr0000gs/T//RtmpSSZc9j/base..agrepl.R 
#> [1] "diff -W 80 -y /var/folders/9z/y1_0kf711fx3lgc68w1k27xr0000gs/T//RtmpSSZc9j/base..grepl.R /var/folders/9z/y1_0kf711fx3lgc68w1k27xr0000gs/T//RtmpSSZc9j/base..agrepl.R"