Get axis label for minorLogTicks

Get axis label for minorLogTicks

getAxisLabel(
  i,
  asValues,
  logAxisType = c("normal", "flip", "pvalue"),
  logBase,
  base_limit = 2,
  offset = 0,
  symmetricZero = (offset > 0),
  ...
)

Arguments

i

numeric axis value

asValues

logical indicating whether the value should be evaluated.

logAxisType

character string with the type of axis values:

"normal": axis values as-is.

"flip": inverted axis values, for example where negative values should be displayed as negative log-transformed values.

"pvalue": for values transformed as -log10(pvalue)

logBase

numeric logarithmic base

base_limit

numeric value indicating the minimum value that should be written as an exponential.

offset

numeric value of offset used for log transformation.

symmetricZero

logical indicating whether negative values should be displayed as negative log-transformed values.

...

additional arguments are ignored.

Value

axis label as character or expression where necessary.

Details

This function is intended to be called internally by jamba::minorLogTicks().

See also