Vectorized isTRUE

Vectorized isTRUE

isTRUEV(x, ...)

Arguments

x

vector

...

additional arguments are ignored

Details

This function applies three criteria to an input vector, to determine if each entry in the vector is TRUE:

It must be class logical.

It must not be NA.

It must evaluate as TRUE.

See also