Show R function arguments jam-style

Show R function arguments jam-style

jargs(x, grepString = NULL, sortVars = TRUE, asList = TRUE,
  useColor = TRUE, verbose = FALSE, ...)

Arguments

x
function or character name of a function.

grepString
NULL, logical, or character grep string used to filter names of function parameters. If logical, it is assumed to be sortVars, and indicates whether to sort the parameter names.

sortVars
logical whether to sort the function parameter names.

asList
logical whether to display one entry per line (TRUE), or display results as a data.frame.

useColor
logical whether to display results in color, if the crayon package is available, and terminal console is capable.

verbose
logical whether to print verbose output.

Details

This function displays R function arguments, organized with one argument per line, and colorized using the crayon package if installed.

It has an optional subset ability, using grepString which filters argument names by pattern matching. This feature is intended to help find the parameter name using a substring, which is particularly helpful for R functions which contain numerous parameters, and which have not obscured them by use of generic function wrapper. Generic functions very often contain no useful parameters, making it difficult to discover required parameters without reading the function documentation from the proper dispatched function and calling package.

It also by default sorts parameters by name, to make it easier to find from among many parameters.

Examples

args(jargs)

#> function (x, grepString = NULL, sortVars = TRUE, asList = TRUE, 
#>     useColor = TRUE, verbose = FALSE, ...) 
#> NULL

jargs(jargs)

#>        ... = ,
#>     asList = TRUE,
#> grepString = NULL,
#>   sortVars = TRUE,
#>   useColor = TRUE,
#>    verbose = FALSE,
#>          x = 

# retrieve parameters involving notes from imageByColors
jargs(imageByColors, "note")

#>       cellnote = NULL,
#>    cexCellnote = 1,
#>   fontCellnote = 1,
#> groupCellnotes = TRUE,
#>    srtCellnote = 0