sort alphanumeric values keeping numeric values in proper order

sort alphanumeric values keeping numeric values in proper order

mixedSort(x, blanksFirst = TRUE, NAlast = TRUE, keepNegative = FALSE,
  keepInfinite = FALSE, keepDecimal = FALSE, ignore.case = TRUE,
  sortByName = FALSE, verbose = FALSE, ...)

Arguments

x
input vector

blanksFirst
logical whether to order blank entries before entries containing a value.

NAlast
logical whether to move NA entries to the end of the sort.

keepNegative
logical whether to keep '-' associated with adjacent numeric values, in order to sort them as negative values.

keepInfinite
logical whether to allow "Inf" to be considered a numeric infinite value.

keepDecimal
logical whether to keep the decimal in numbers, sorting as a true number and not as a version number. By default keepDecimal=FALSE, which means "v1.200" should be ordered before "v1.30". When keepDecimal=TRUE, the numeric sort considers only "1.2" and "1.3" and sorts in that order.

ignore.case
logical whether to ignore uppercase and lowercase characters when defining the sort order.

sortByName
logical whether to sort the vector x by names(x) instead of sorting by x itself.

verbose
logical whether to print verbose output.

...
additional parameters are sent to mixedOrder.

Value

vector of values from x, ordered by alphanumeric logic.

Details

This function is a refactor of the mixedsort function from the gtools package. It was extended to make it faster, and to handle special cases slightly differently. It was driven by some need to sort gene symbols, and miRNA symbols in numeric order, for example:

test set:

miR-12,miR-1,miR-122,miR-1b,mir-1a

gtools::mixedsort:

miR-122,miR-12,miR-1,miR-1a,mir-1b

mixedSort:

miR-1,miR-1a,miR-1b,miR-12,miR-122

The function does not by default recognize negative numbers as negative, instead it treats '-' as a delimiter, unless keepNegative=TRUE.

This function also attempts to maintain '.' as part of a decimal number, which can be problematic when sorting IP addresses, for example.

This function is really just a wrapper function for mixedOrder which does the work of defining a proper order.

See also

mixedSort, mixedorder, mixedsort

Examples

x <- c("miR-12","miR-1","miR-122","miR-1b", "miR-1a", "miR-2");
sort(x);

#> [1] "miR-1"   "miR-12"  "miR-122" "miR-1a"  "miR-1b"  "miR-2"  

mixedSort(x);

#> [1] "miR-1"   "miR-1a"  "miR-1b"  "miR-2"   "miR-12"  "miR-122"