order alphanumeric values from a list

order alphanumeric values from a list

mmixedOrder(..., na.last = TRUE, decreasing = FALSE, verbose = FALSE,
  ignore.case = TRUE, matrixAsDF = TRUE)

Arguments

...
parameters treated as a list of vectors to be ordered in proper order, based upon the mechanism by order.

na.last, decreasing, verbose, ignore.case
parameters sent to mixedOrder.

matrixAsDF
logical if ... supplies only one matrix object, whether to convert it to data.frame, the coerce to a list, before processing. By default, in the event only one matrix object is supplied, this conversion is performed, in order to define a sort order based upon each column in order.

Value

integer vector of row orders

Details

This function is a minor extension to mixedOrder in that it accepts list input, similar to how order operates. This function is mainly useful when sorting something like a data.frame, where ties in column 1 should be maintained then broken by non-equal values in column 2.

In fact, mixedSortDF calls this mmixedOrder function, in order to sort a data.frame properly by column.

See mixedOrder and mixedSort for a better description of how the sort order logic operates.

See also