assign unique names for a vector

assign unique names for a vector

nameVector(x, y = NULL, makeNamesFunc = makeNames, ...)

Arguments

x
vector input, or data.frame with two columns, the second column is used to name values in the first column.

y
NULL or character vector of names. If NULL then x is used. Note that y is recycled to the length of x, prior to being sent to the makeNamesFunc.

makeNamesFunc
function to make names unique, by default makeNames which ensures names are unique.

...
passed to makeNamesFunc

Details

This function assigns unique names to a vector, if necessary it runs makeNames to create unique names. It differs from setNames in that it ensures names are unique, and when no names are supplied, it uses the vector itself to define names. It is helpful to run this function inside an lapply function call, which by default maintains names, but does not assign names if the input data did not already have them.

When used with a data.frame, it is particularly convenient to pull out a named vector of values. For example, log2 fold changes by gene, where the gene symbols are the name of the vector. nameVector(genedata[,c("Gene","log2FC")])

Examples

# it generally just creates names from the vector values
nameVector(LETTERS[1:5]);

#>   A   B   C   D   E 
#> "A" "B" "C" "D" "E" 

# if values are replicated, the makeNames() function makes them unique
V <- rep(LETTERS[1:5], c(1,2,3,4,5));
nameVector(V);

#>    A B_v1 B_v2 C_v1 C_v2 C_v3 D_v1 D_v2 D_v3 D_v4 E_v1 E_v2 E_v3 E_v4 E_v5 
#>  "A"  "B"  "B"  "C"  "C"  "C"  "D"  "D"  "D"  "D"  "E"  "E"  "E"  "E"  "E" 

# for a two-column data.frame, it creates a named vector using
# the values in the first column, and names in the second column.
df <- data.frame(seq_along(V), V);
df;

#>    seq_along.V. V
#> 1             1 A
#> 2             2 B
#> 3             3 B
#> 4             4 C
#> 5             5 C
#> 6             6 C
#> 7             7 D
#> 8             8 D
#> 9             9 D
#> 10           10 D
#> 11           11 E
#> 12           12 E
#> 13           13 E
#> 14           14 E
#> 15           15 E

nameVector(df);

#>    A B_v1 B_v2 C_v1 C_v2 C_v3 D_v1 D_v2 D_v3 D_v4 E_v1 E_v2 E_v3 E_v4 E_v5 
#>    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15