Import one or more data.frame from Excel xlsx format

Import one or more data.frame from Excel xlsx format

readOpenxlsx(
  xlsx,
  sheet = NULL,
  startRow = 1,
  rows = NULL,
  check.names = FALSE,
  check_header = TRUE,
  check_header_n = 10,
  verbose = FALSE,
  ...
)

Arguments

xlsx

character path to an Excel file in xlsx format, compatible with openxlsx::read.xlsx().

sheet

one of NULL, character, or integer vector, where: sheet=NULL will import every sheet; character is a vector of sheet names; and integer is a vector of sheet index values. The sheet names are determined with openxlsx::getSheetNames().

startRow

integer indicating the row number to start importing each sheet. Note that startRow can be a vector with length length(sheet), to specify the startRow for each sheet.

rows

integer vector indicating specific rows to import for each sheet. To specify different rows for each sheet, supply rows as a list of integer vectors.

check.names

logical indicating whether to call make.names() on the colnames of each data.frame.

check_header

logical indicating whether to test for presence of header rows, which are not column headers. When check_header=TRUE, this method simply tests for the presence of rows that have ncol different than the remaining rows of data in the given sheet. When header rows are detected, the values are assigned to column dimnames of the data.frame.

check_header_n

integer number of rows to test for header rows, only used when check_header=TRUE.

verbose

logical indicating whether to print verbose output.

...

additional arguments are passed to openxlsx::read.xlsx().

Details

This function is equivalent to openxlsx::read.xlsx() with a few minor additions:

It returns a list of data.frame objects, one per sheet.

It properly reads the colnames with check.names=FALSE.

By default this function returns every sheet for a given xlsx file.

See also