Reload Rmarkdown cache

Reload Rmarkdown cache in the order files were created, into an R environment

Usage

reload_rmarkdown_cache(
  dir = ".",
  maxnum = 1000,
  max_cache_name = NULL,
  envir = globalenv(),
  file_sort = c("globals", "objects", "ctime", "mtime"),
  preferred_load_types = c("lazyLoad", "load"),
  dryrun = FALSE,
  verbose = TRUE,
  ...
)

Arguments

dir

character path to the directory that contains Rmarkdown cache files. Each file is recognized by the file extension ".rdx".

maxnum

integer indicating the maximum number of cache files to re-load, in order.

max_cache_name

character optional string indicating the name of an Rmarkdown cache chunk where this function will stop loading cache data. All cache files after this point will not be loaded. This option is intended to help recreate the data available to a particular Rmarkdown chunk during processing.

envir

environment where cache data will be loaded.

file_sort

character string indicating how to sort cache files to place them in proper order for re-loading. The default is to use one RMarkdown cache index file, which should accurately reflect only RMarkdown chunks present in the .Rmd file during rendering, and in the order they appear in that file. Using "mtime" or "ctime" below will sort files by the modification or creation time, respectively, and is less accurate, but often sufficient for most purposes. It would only be advised if for some reason the "__globals" or "__objects" files are not present.

"globals" uses the "__globals" file in the cache directory.

"objects" uses the "__objects" file in the cache directory.

ctime sorts by file creation time, default

mtime sorts by file modification time

preferred_load_types

character string indicating the preferred load mechanism to use. By default either option is permitted, depending upon the files present: it will use lazyLoad() if .rdx/.rdb files are present, otherwise it falls back to using load() for .RData files.

"lazyLoad" will try to use lazyLoad() to load .rdx/.rdb files

"load" will try to use load() to load .RData files

Remove "lazyLoad" to prevent lazy-loading of cached objects.

dryrun

logical indicating whether to perform a dry-run, which prints messages but does not process the data.

verbose

logical indicating whether to print verbose output. Note that this variable is not passed along to load(), since it is inconsistent with lazyLoad().

...

additional arguments are passed to lazyLoad() or load() as relevant to the method used to re-load the cache object data.

Value

this function does not return data, but instead is called for the by-product of loading data into the given envir environment.

Details

This function is intended to help re-load Rmarkdown cache files created during the processing/rendering of an Rmarkdown file.

By default, all cached R objects are loaded into the global environment globalenv(), However, it is recommended that envir is used to define a new environment into which the cached session is loaded.

cache_env <- new.env()
reload_rmarkdown_cache(cachedir, envir=cache_env)

From then on, the cached data objects can be seen with ls(cache_env) and retrieved with get("objectname", envir=cache_env).

If supplied with maxnum or max_cache_name then the cache will be loaded only up to this point, and not beyond. The recommended method to determine the cache is to use dryrun=TRUE to view all sections, then to choose the integer number, or character name to define the maximum chunk to load.

See also