Define visible text color

Given a vector or colors, define a contrasting color for text, typically using either white or black.

setTextContrastColor(color, hclCutoff = 73, rgbCutoff = 127,
  colorModel = c("hcl", "rgb"), useGrey = 0, keepAlpha = FALSE, ...)

Arguments

color
character vector with one or more R-compatible colors.

hclCutoff
numeric threshold above which a color is judged to be bright, therefore requiring a dark text color. This comparison uses the L value from the col2hcl function, which scales colors from 1 to 100.

rgbCutoff
numeric threshold above which a color is judged to be bright, therefore requiring a dark text color. The mean r,g,b value is used.

colorModel
Either 'hcl' or 'rgb' to indicate how the colors will be judged for overall brightness. The 'hcl' method uses the L value, which more reliably represents overall visible lightness.

useGrey
numeric threshold used to define dark and bright text colors, using the R greyscale gradient from 0 to 100. useGrey=10 implies 'grey10' and 'grey90' for the contrasting text colors. useGrey=15 is useful if labels may also overlap white or black space, since the text will never be fully white or black.

keepAlpha
logical indicates whether the input color alpha transparency should be maintained in the text color. By default, text alpha is not maintained, and instead is set to alpha=1, fully opaque.

See also

jamba color manipulation functions: hsv2col, col2hsv,col2hcl, rgb2col, alpha2col,col2alpha, makeColorDarker, color2gradient, setTextContrastColor, showColors

Examples

setTextContrastColor(c("red","yellow","lightblue","blue4"))

#>         red      yellow   lightblue       blue4 
#> "#FFFFFFFF" "#000000FF" "#000000FF" "#FFFFFFFF"