Smooth scatter plot, Jam style

Produce smooth scatter plot, a helper function called by plotSmoothScatter.

smoothScatterJam(x, y = NULL, nbin = 128, bandwidth,
  colramp = colorRampPalette(c("white", blues9)), nrpoints = 100,
  pch = ".", cex = 1, col = "black", transformation = function(x)
  x^0.25, postPlotHook = box, xlab = NULL, ylab = NULL, xlim, ylim,
  add = FALSE, xaxs = par("xaxs"), yaxs = par("yaxs"),
  xaxt = par("xaxt"), yaxt = par("yaxt"), useRaster = NULL, ...)

Arguments

x
numeric vector, or data matrix with two or more columns.

y
numeric vector, or if data is supplied via x as a matrix, y is NULL.

nbin
integer number of bins to use when converting the kernel density result (which uses bandwidthN above) into a usable image. For example, nbin=123 is the default used by smoothScatter, however the plotSmoothScatter function default is higher (256).

bandwidth
numeric vector used to define the y- and x-axis bandwidths, respectively, for the hidden but very useful grDevices:::.smoothScatterCalcDensity function, which calculates the underlying 2-dimensional kernel density of data points. This parameter is also why the wrapper function plotSmoothScatter was created, in order to avoid ever having to define this parameter directly.

nrpoints
integer number of outlier datapoints to display, as defined by the hidden but very useful grDevices:::.smoothScatterCalcDensity function. The base link[graphics]{smoothScatter} function plots 100 such points, perhaps to overcome the default over-smoothing of data which results in large areas not displaying density. The default here is zero, since the new default bandwidthN parameter typically already indicates these points.

transformation
function which converts point density to a number, typically related to square root or cube root transformation.

postPlotHook
is NULL for no post-plot hook, or a function which is called after producing the image plot. By default it is simply used to draw a box around the image, but could be used to layer additional information atop the image plot, for example contours, labels, etc.

xlab
character x-axis label

ylab
character y-axis label

xlim
numeric x-axis range for the plot

ylim
numeric y-axis range for the plot

add
logical whether to add to an existing active R plot, or create a new plot window.

xaxs
character value compatible with par(xaxs), mainly useful for suppressing the x-axis, in order to produce a custom x-axis range, most useful to restrict the axis range expansion done by R by default.

yaxs
character value compatible with par(yaxs), mainly useful for suppressing the y-axis, in order to produce a custom y-axis range, most useful to restrict the axis range expansion done by R by default.

xaxt
character value compatible with par(xaxt), mainly useful for suppressing the x-axis, in order to produce a custom x-axis by other mechanisms, e.g. log-scaled x-axis tick marks.

yaxt
character value compatible with par(yaxt), mainly useful for suppressing the y-axis, in order to produce a custom y-axis by other mechanisms, e.g. log-scaled y-axis tick marks.

useRaster
NULL or logical, indicating whether to invoke rasterImage to produce a raster image. If NULL, it determines whether to produce a raster image within the imageDefault function, which checks the options using getOption("preferRaster", FALSE) to determine among other things, whether the user prefers raster images, and if the dev.capabilities supports raster.

Details

This function is only necessary in order to override the smoothScatter function which calls image.default(). Instead, this function calls imageDefault which is required in order to utilize custom raster image scaling, particularly important when the x- and y-axis ranges are not similar, e.g. where the x-axis spans 10 units, but the y-axis spans 10,000 units. The bulk of this function and its parameters are simply copied from smoothScatter for consistency with the parent function, with due credit and respect to its authors.

See also

plotSmoothScatter, imageDefault, smoothScatter