print dimensions of nested list objects

print dimensions of list nested objects, such as a list of lists

ssdim(
  x,
  includeClass = TRUE,
  doFormat = FALSE,
  big.mark = ",",
  verbose = FALSE,
  ...
)

Arguments

x
one of several recognized object classes: an S3 object inheriting from class "list", including a nested list of lists or simple list; an S4 object in which case it used slotNames(x) to traverse the object structure; an "environment" object, in which case ls(envir=x) is used to traverse the object structure. When the object is S4 that inherits "List" from the S4Vectors package, it will attempt to use the proper subset functions from S4Vectors via names(x), but that process only works properly if the S4Vectors package is previously loaded, otherwise it reverts to using slotNames(x).

includeClass
boolean passed to sdim(), indicating whether to print the class of each element in the input x object. Note that for S4 objects, each element will be the object returned for each of slotNames(x).

doFormat
boolean passed to sdim(), indicating whether to format the dimensions using format(...,big.mark=","), which is mainly useful for extremely large dimensions. This parameter should probably become more broadly useful and respectful for different locales.

big.mark
character passed to sdim(), value used when doFormat=TRUE, used in the call to format(...,big.mark).

verbose
logical whether to print verbose output

...
additional parameters are ignored.

Value

list of data.frame, each row indicates the dimensions of each element in the input list. When includeClass is TRUE it will include a column class which indicates the class of each list element. When the input list contains arrays with more than two dimensions, the first two dimensions are named "rows" and "columns" with additional dimensions named "dim3" and so on. Any list element with fewer than that many dimensions will only have values populated to the relevant dimensions, for example a character vector will only populate the length.

Details

This function iteratively calls sdim(x) on each list element, which can be helpful for some more complicated object formats. In future, the intent is to recurse through a nested list structure, but currently this process only recurses two steps. Attempts to recurse deeper sometimes results in printing much more detail than originally intended.

See also

Examples

## Simple list
L <- list(LETTERS=LETTERS,
   letters=letters,
   lettersDF=data.frame(LETTERS, letters));
sdim(L);

#>           rows cols      class
#> LETTERS     26       character
#> letters     26       character
#> lettersDF   26    2 data.frame

## list of lists
L2 <- list(List1=L,
   List2=rev(L));
## first level of detail
sdim(L2);

#>       rows class
#> List1    3  list
#> List2    3  list

## second level of detail
ssdim(L2);

#> $List1
#>           rows cols      class
#> LETTERS     26       character
#> letters     26       character
#> lettersDF   26    2 data.frame
#> 
#> $List2
#>           rows cols      class
#> lettersDF   26    2 data.frame
#> letters     26       character
#> LETTERS     26       character
#> 

## Perform the same steps using an environment
E1 <- new.env();
assign("L", envir=E1, value=L);
sdim(E1);

#>   rows class
#> L    3  list

ssdim(E1);

#> $L
#>           rows cols      class
#> LETTERS     26       character
#> letters     26       character
#> lettersDF   26    2 data.frame
#> 

assign("L2", envir=E1, value=L2);
ssdim(E1);

#> $L
#>           rows cols      class
#> LETTERS     26       character
#> letters     26       character
#> lettersDF   26    2 data.frame
#> 
#> $L2
#>       rows class
#> List1    3  list
#> List2    3  list
#> 

rm(E1);