Remove alpha transparency from colors

Remove alpha transparency from colors

unalpha(x, ...)

Value

character vector of R colors in hex format.

Details

This function simply removes the alpha transparency from R colors, returned in hex format, for example "#FF0000FF" becomes "#FF0000", or "blue" becomes "#0000FF".

It also silently converts R color names to hex format, where applicable.

See also