convert numeric value to size

convert numeric value to size

asSize(
  x,
  humanFriendly = TRUE,
  digits = 3,
  abbreviateUnits = TRUE,
  unitType = "bytes",
  unitAbbrev = gsub("^(.).*$", "\\1", unitType),
  kiloSize = 1024,
  sep = " ",
  ...
)

Arguments

x
numeric vector

humanFriendly
logical, currently only TRUE is accepted, whether to include human-friendly units to the output.

digits
integer number of digits used by format when formatting the number to create a character string

abbreviateUnits
logical whether to print abbreviated units, for example using k, M, G, T, P instead of kilo, mega, Giga, Tera, Peta, respectively.

unitType
character string indicating the base unit of measure, by default "bytes". Note that trailing "s" is removed when the number is singular.

unitAbbrev
character string indicating an abbreviated base unit, by default it uses the first character from unitType.

kiloSize
numeric number of base units when converting from one base unit, to one "kilo" base unit. For file sizes, this value is 1024, but for other purposes this value may be 1000.

sep
delimiter used between the numeric value and the unit.

...
other parameters passed to format.

Value

character vector representing human-friendly sizes.

Details

This function converts numeric input to computer size in bytes, using the most appropriate human-friendly units as possible. For example, it will use kilobytes (kb), megabytes (Mb), gigabytes (Gb), etc. as necessary.

This function by default assumes 1,024 bytes per kilobyte, but can be made to use 1000, for example in order to represent genome base size, which is represented in kilobases (kb), megabases (Mb), gigabases (Gb), etc. Simply define unitType="bases" in this scenario.

TODO: Allow units to be specified as a function argument, for example "million" and "billion" instead of "Giga" and "Tera".

See also

Examples

asSize(c(1, 10,2010,22000,52200))

#> [1] "1 byte"   "10 bytes" "2 kb"     "21 kb"    "51 kb"   

#> "1 byte"   "10 bytes" "2 kb"     "21 kb"    "51 kb"

# demonstration of straight numeric units
asSize(c(1, 100, 1000, 10000), unitType="", kiloSize=100)

#> [1] "1"    "1 k"  "10 k" "1 M"