convert R color to HCL color matrix

convert R color to HCL color matrix

col2hcl(
  x,
  maxColorValue = 255,
  model = getOption("jam.model", c("hcl", "polarLUV", "polarLAB")),
  ...
)

Arguments

x

R compatible color, either a color name, or hex value, or a mixture of the two. Any value compatible with col2rgb.

model

the color model to use, from "polarLUV" for the standard R conventional HCL, and "polarLAB" which uses the LAB-based HCL values.

...

additional arguments are ignored.

maxValue

numeric maximum value to return, useful when the downstream alpha range should be 255. By default maxValue=1 is returned.

Details

This function takes an R color and converts to an HCL matrix, using the colorspace package, and RGB and polarLUV functions. It is also used to maintain alpha transparency, to enable interconversion via other color manipulation functions as well.

When model="hcl" this function uses farver::decode_colour() and bypasses colorspace. In future the colorspace dependency will likely be removed in favor of using farver. In any event, model="hcl" is equivalent to using model="polarLUV" and fixup=TRUE, except that it should be much faster.

See also

Examples

col2hcl("#FF000044")
#>              [,1]
#> H      12.1739887
#> C     179.0413601
#> L      53.2407941
#> alpha   0.2666667