Handle function arguments as text

Handles a list or list of lists, converting to human-readable text format

handleArgsText(argTextA, name = "", col1 = "mediumpurple2",
  col2 = "mediumaquamarine", colT = "dodgerblue3", colF = "red1",
  colNULL = "grey60", lightMode = checkLightMode(),
  adjustRgb = getOption("jam.adjustRgb"), Crange = NULL,
  Lrange = NULL, indent = "", useCollapseList = ",\n      ",
  useCollapseBase = ", ", level = 1, debug = 0, useColor = TRUE,
  verbose = FALSE, ...)

Details

This function is a rare non-exported function intended to be called by jargs(), but separated in order to help isolate the logical steps required.

See also