Vectorized isTRUE

Vectorized isTRUE

Usage

isTRUEV(x, ...)

Arguments

x

vector

...

additional arguments are ignored

Value

logical vector with length matching x.

Details

This function applies three criteria to an input vector, to determine if each entry in the vector is TRUE:

It must be class logical.

It must not be NA.

It must evaluate as TRUE.

See also

Examples

isTRUEV(c(TRUE, FALSE, NA, TRUE))
#> [1]  TRUE FALSE FALSE  TRUE