Show R function arguments jam-style

Show R function arguments jam-style

jargs(
  x,
  grepString = NULL,
  sortVars = FALSE,
  asList = TRUE,
  useColor = TRUE,
  lightMode = NULL,
  Crange = getOption("jam.Crange"),
  Lrange = getOption("jam.Lrange"),
  adjustRgb = getOption("jam.adjustRgb"),
  useCollapseBase = ", ",
  verbose = FALSE,
  debug = 0,
  ...
)

Arguments

x

function or character name of a function.

grepString

NULL, logical, or character grep string used to filter names of function parameters. If logical, it is assumed to be sortVars, and indicates whether to sort the parameter names.

sortVars

logical whether to sort the function parameter names.

asList

logical whether to display one entry per line (TRUE), or display results as a data.frame.

useColor

logical whether to display results in color, if the crayon package is available, and terminal console is capable.

lightMode

boolean or NULL, indicating whether the text background color is light, thus imposing a maximum brightness for colors displayed. It use lightMode if defined by the function caller, otherwise it will use options("jam.lightMode") if defined, lastly it will attempt to detect whether running inside Rstudio by checking the environment variable "RSTUDIO", and if so it will assume lightMode==TRUE.

Crange

numeric range of chroma values, ranging between 0 and 100. When NULL, default values will be assigned to Crange by setCLranges().

Lrange

numeric range of luminance values, ranging between 0 and 100. When NULL, default values will be assigned to Lrange by setCLranges().

adjustRgb

numeric value adjustment used during the conversion of RGB colors to ANSI colors, which is inherently lossy. If not defined, it uses the default returned by setCLranges() which itself uses getOption("jam.adjustRgb") with default=0. In order to boost color contrast, an alternate value of -0.1 is suggested.

useCollapseBase

character string used to combine multiple parameter values.

verbose

logical whether to print verbose output.

debug

integer value, greater than 0 will cause debug-type verbose output, useful because parameters are hard!

Details

This function displays R function arguments, organized with one argument per line, and colorized using the crayon package if installed.

It has an optional subset ability, using grepString which filters argument names by pattern matching. This feature is intended to help find the parameter name using a substring, which is particularly helpful for R functions which contain numerous parameters, and which have not obscured them by use of generic function wrapper. Generic functions very often contain no useful parameters, making it difficult to discover required parameters without reading the function documentation from the proper dispatched function and calling package.

It also by default sorts parameters by name, to make it easier to find from among many parameters.

See also

Examples

args(jargs)
#> function (x, grepString = NULL, sortVars = FALSE, asList = TRUE, 
#>     useColor = TRUE, lightMode = NULL, Crange = getOption("jam.Crange"), 
#>     Lrange = getOption("jam.Lrange"), adjustRgb = getOption("jam.adjustRgb"), 
#>     useCollapseBase = ", ", verbose = FALSE, debug = 0, ...) 
#> NULL
jargs(jargs)
#>               x = ,
#>      grepString = NULL,
#>        sortVars = FALSE,
#>          asList = TRUE,
#>        useColor = TRUE,
#>       lightMode = NULL,
#>          Crange = getOption("jam.Crange"),
#>          Lrange = getOption("jam.Lrange"),
#>       adjustRgb = getOption("jam.adjustRgb"),
#> useCollapseBase = ", ",
#>         verbose = FALSE,
#>           debug = 0,
#>             ... = 

# retrieve parameters involving notes from imageByColors
jargs(imageByColors, "note")
#>       cellnote = NULL,
#>    cexCellnote = 1,
#>    srtCellnote = 0,
#>   fontCellnote = 1,
#> groupCellnotes = TRUE