Convert list of vectors to data.frame with item, value, name

Convert list of vectors to data.frame with item, value, name

list2df(x, makeUnique = TRUE, useVectorNames = TRUE, ...)

Arguments

x

list of vectors

makeUnique

logical indicating whether the data.frame should contain unique rows.

useVectorNames

logical indicating whether vector names should be included in the data.frame, if they exist.

...

additional arguments are ignored.

Details

This function converts a list of vectors to a tall data.frame with colnames item to indicate the list name, value to indicate the vector value, and name to indicate the vector name if useVectorNames=TRUE and if names exist.

See also