vectorized make_styles for crayon output

vectorized make_styles for crayon output

make_styles(style, text, bg = FALSE, grey = FALSE,
  colors = num_colors(), Cgrey = getOption("jam.Cgrey"),
  lightMode = checkLightMode(), Crange = getOption("jam.Crange"),
  Lrange = getOption("jam.Crange"),
  adjustRgb = getOption("jam.adjustRgb"), adjustPower = 1.5,
  fixYellow = TRUE, colorTransparent = "grey45", alphaPower = 2,
  setOptions = FALSE, verbose = FALSE, ...)

Arguments

style
vector of one or more styles

text
vector of one or more character values

bg
NULL or a vector of one or more background styles

colors
integer number of colors allowed for console output

Cgrey
numeric chroma (C) value, which defines grey colors at or below this chroma. Any colors at or below the grey cutoff will have use ANSI greyscale coloring. To disable, set Cgrey=-1.

lightMode
boolean indicating whether the background color is light (TRUE is bright), or dark (FALSE is dark.) By default it calls checkLightMode() which queries getOption("lightMode").

Crange
numeric range of chroma values, ranging between 0 and 100. When NULL, default values will be assigned to Crange. When supplied, range(Crange) is used.

Lrange
numeric range of luminance values, ranging between 0 and 100. When NULL, default values will be assigned to Lrange. When supplied, range(Lrange) is used.

adjustRgb
numeric value adjustment used during the conversion of RGB colors to ANSI colors, which is inherently lossy. If not defined, it uses the default returned by setCLranges() which itself uses getOption("jam.adjustRgb") with default=0. In order to boost color contrast, an alternate value of -0.1 is suggested.

adjustPower
numeric adjustment power factor

fixYellow
boolean indicating whether to "fix" the darkening of yellow, which otherwise turns to green. Instead, since JAM can, JAM will make the yellow slightly more golden before darkening. This change only affects color hues between 80 and 90. This argument is passed to applyCLrange().

colorTransparent
color used to substitute for "transparent" which a valid R color, but not a valid color for the crayon package.

alphaPower
numeric value, used to adjust the RGB values for alpha values less than 255, by raising the ratio to 1/alphaPower, which takes the ratio of square roots. alphaPower=100 for minimal adjustment.

setOptions
logical whether to update Crange and Lrange options during the subsequent call to setCLranges(). By default, FALSE prevents modifying the global settings, in favor of using a direct call to setCLranges() itself, outside of make_styles() to make such updates.

verbose
logical whether to print verbose output

...
additional parameters are ignored

satCutoff
numeric cutoff for color saturation, below which a color is considered "grey" and the ANSI greyscale color set is used.

Value

Vector with the same length as text input vector, where entries are surrounded by the relevant encoding consistent with the style defined at input. In short, a character vector as input, a colorized character vector as output.

Details

This function is essentially a vectorized version of make_style in order to style a vector of character strings with a vector of foreground and background styles.

See also