order alphanumeric values keeping numeric values in proper order

order alphanumeric values keeping numeric values in proper order

mixedOrder(x, ..., blanksFirst = TRUE, NAlast = TRUE,
  keepNegative = FALSE, keepInfinite = FALSE, keepDecimal = FALSE,
  verbose = FALSE, ignore.case = TRUE, useCaseTiebreak = TRUE,
  returnDebug = FALSE, returnType = c("order", "rank"))

Arguments

x
input vector

...
additional parameters are sent to mixedOrder.

blanksFirst
logical whether to order blank entries before entries containing a value.

NAlast
logical whether to move NA entries to the end of the sort.

keepNegative
logical whether to keep '-' associated with adjacent numeric values, in order to sort them as negative values.

keepInfinite
logical whether to allow "Inf" to be considered a numeric infinite value.

keepDecimal
logical whether to keep the decimal in numbers, sorting as a true number and not as a version number. By default keepDecimal=FALSE, which means "v1.200" should be ordered before "v1.30". When keepDecimal=TRUE, the numeric sort considers only "1.2" and "1.3" and sorts in that order.

verbose
logical whether to print verbose output.

ignore.case
logical whether to ignore uppercase and lowercase characters when defining the sort order.

sortByName
logical whether to sort the vector x by names(x) instead of sorting by x itself.

Value

integer vector of orders derived from x

Details

This function is a refactor of the gtools::mixedorder function from the gtools package. It was extended to make it faster, and to handle special cases slightly differently. It was driven by some need to sort gene symbols, and miRNA symbols in numeric order, for example:

test set:

miR-12,miR-1,miR-122,miR-1b,miR-1a,miR-2

sort:

miR-1,miR-12,miR-122,miR-1a,miR-1b,miR-2

gtools::mixedsort:

miR-122,miR-12,miR-2,miR-1,miR-1a,miR-1b

mixedSort:

miR-1,miR-1a,miR-1b,miR-2,miR-12,miR-122

The function does not by default recognize negative numbers as negative, instead it treats '-' as a delimiter, unless keepNegative=TRUE.

This function also attempts to maintain '.' as part of a decimal number, which can be problematic when sorting IP addresses, for example.

This function is also available as a sort function mixedSort.

See also

Examples

x <- c("miR-12","miR-1","miR-122","miR-1b", "miR-1a","miR-2");
mixedOrder(x);

#> [1] 2 5 4 6 1 3

x[mixedOrder(x)];

#> [1] "miR-1"   "miR-1a"  "miR-1b"  "miR-2"   "miR-12"  "miR-122"

mixedSort(x);

#> [1] "miR-1"   "miR-1a"  "miR-1b"  "miR-2"   "miR-12"  "miR-122"

order(x);

#> [1] 2 1 3 5 4 6

x[order(x)];

#> [1] "miR-1"   "miR-12"  "miR-122" "miR-1a"  "miR-1b"  "miR-2"  

sort(x);

#> [1] "miR-1"   "miR-12"  "miR-122" "miR-1a"  "miR-1b"  "miR-2"