sort data.frame keeping numeric values in proper order

sort data.frame keeping numeric values in proper order

mixedSortDF(df, byCols = seq_len(ncol(df)), na.last = TRUE,
  decreasing = NULL, useRownames = TRUE, verbose = FALSE, ...)

Arguments

df
data.frame input

byCols
integer vector referring to the order of columns to be used by mmixedOrder to order the data.frame. Note that negative values will reverse the sort order for the corresponding column number. byCols can also be a character vector of values in colnames(df), optionally including prefix "-" to reverse the sort. Note that the parameter decreasing can also be used to specify columns to have reverse sort.

na.last
logical whether NA values should be ranked last

decreasing
NULL or a logical vector indicating which columns in byCols should be sorted in decreasing order. By default, the sign(byCols) is used to define the sort order of each column, but it can be explicitly overridden with this decreasing parameter.

useRownames
logical whether to use rownames(df) as a last tiebreaker in the overall rank ordering. This parameter has the primary effect of assuring a reproducible result, provided the rownames are consistently defined, or if rownames are actually row numbers.

verbose
logical whether to print verbose output

...
additional arguments passed to mmixedOrder(...)

Value

data.frame whose rows are ordered using mmixedOrder

Details

This function is a wrapper around mmixedOrder so it operates on data.frame columns in the proper order, using logic similar that used by order when operating on data.frames. The sort order logic is fully described in mixedSort and mixedOrder.

See also

Examples

# start with a vector of miRNA names
x <- c("miR-12","miR-1","miR-122","miR-1b", "miR-1a","miR-2");
# add some arbitrary group information
g <- rep(c("Air", "Treatment", "Control"), 2);
# create a data.frame
df <- data.frame(group=g, miRNA=x, stringsAsFactors=FALSE);

# sort the data.frame by each column
mixedSortDF(df);

#>       group   miRNA
#> 4       Air  miR-1b
#> 1       Air  miR-12
#> 6   Control   miR-2
#> 3   Control miR-122
#> 2 Treatment   miR-1
#> 5 Treatment  miR-1a

# mixedSort respects factor order, so reorder the factor levels
# to demonstrate. "Control" should come first, for example.'
gf <- factor(g, levels=c("Control","Air", "Treatment"));
df2 <- data.frame(groupfactor=gf, miRNA=x, stringsAsFactors=FALSE);

# now the sort properly keeps the group factor levels in order,
# which also sorting the miRNA names in their proper order.
mixedSortDF(df2);

#>   groupfactor   miRNA
#> 6     Control   miR-2
#> 3     Control miR-122
#> 4         Air  miR-1b
#> 1         Air  miR-12
#> 2   Treatment   miR-1
#> 5   Treatment  miR-1a

x <- data.frame(l1=letters[1:10],
   l2=rep(letters[1:2+10], 5),
   L1=LETTERS[1:10],
   L2=rep(LETTERS[1:2+20], each=5));