Paste data.frame rows into an ordered factor

Paste data.frame rows into an ordered factor

pasteByRowOrdered(x, sep = "_", na.rm = TRUE, condenseBlanks = TRUE,
  includeNames = FALSE, ...)

Arguments

x
data.frame

sep
character separator to use between columns

na.rm
logical whether to remove NA values, or include them as "NA"

condenseBlanks
logical whether to condense blank or empty values without including an extra delimiter between columns.

includeNames
logical whether to include the colname delimited prior to the value, using sepName as the delimiter.

...
additional arguments are passed to jamba::pasteByRow().

verbose
logical whether to print verbose output.

Details

This function is an extension to jamba::pasteByRow() which pastes rows from a data.frame into a character vector. This function orders the resulting character vector after ordering the data.frame with jamba::mixedSortDF(), which maintains the proper factor level ordering of each column.

Note that the order of columns in the data.frame is the sort order of factors. In future, the column sort order may become an optional parameter.

Note also that jamba::mixedSortDF() uses jamba::mixedSort() which itself performs alphanumeric sort in order to keep values in proper numeric order where possible.

See also