print colorized output to R console

print colorized output to R console

printDebug(
  ...,
  fgText = NULL,
  bgText = NULL,
  fgTime = "cyan",
  timeStamp = TRUE,
  comment = TRUE,
  formatNumbers = TRUE,
  trim = TRUE,
  digits = NULL,
  nsmall = 0L,
  justify = "left",
  big.mark = "",
  small.mark = "",
  zero.print = NULL,
  width = NULL,
  doColor = NULL,
  splitComments = FALSE,
  collapse = "",
  sep = ",",
  detectColors = TRUE,
  darkFactor = c(1, 1.5),
  sFactor = c(1, 1.5),
  lightMode = checkLightMode(),
  Crange = NULL,
  Lrange = NULL,
  removeNA = FALSE,
  replaceNULL = NULL,
  adjustRgb = getOption("jam.adjustRgb"),
  byLine = FALSE,
  verbose = FALSE,
  indent = "",
  keepNA = TRUE,
  file = "",
  append = TRUE,
  invert = FALSE,
  htmlOut = FALSE,
  x
)

Arguments

...
text to be printed to the R console.

fgText
vector of R compatible colors, or a list of vectors of R compatible colors, to define the foreground colors. In the latter case, each vector is applied to each list item from '...'

bgText
vector of R compatible colors, or a list of vectors, to define the background color.

fgTime
character R color to colorize the time

timeStamp
logical whether to include a time stamp in output

comment
logical whether to prefix output with '##' as a comment

formatNumbers
logical whether to format numbers using format which controls the number of digits displayed.

trim, digits, nsmall, justify, big.mark, small.mark, zero.print, width
parameters sent to the format function.

doColor
NULL or logical indicating whether to colorize output. If NULL it detects whether the crayon package is available and console color is enabled.

splitComments
logical whether to color each element independently without light-dark alternating pattern.

collapse
character collapse string used to separate list items, by default "" so text separation is expected in the input data.

sep
character separator used to separate vector elements, when a list items contains a vector.

detectColors
logical whether to detect and potentially try to correct console color capabilities.

darkFactor
numeric darkness to apply to alternative vector values when using alternating light-dark color shading.

sFactor
numeric color saturation to apply to alternative vector values when using alternating light-dark color shading.

lightMode
boolean or NULL, indicating whether the text background color is light, thus imposing a maximum brightness for colors displayed. It use lightMode if defined by the function caller, otherwise it will use options("jam.lightMode") if defined, lastly it will attempt to detect whether running inside Rstudio by checking the environment variable "RSTUDIO", and if so it will assign lightMode TRUE.

Crange
numeric range of chroma values, ranging between 0 and 100. When NULL, default values will be assigned to Crange by setCLranges().

Lrange
numeric range of luminance values, ranging between 0 and 100. When NULL, default values will be assigned to Lrange by setCLranges().

removeNA
logical whether to remove NA values and not print to the console.

replaceNULL
character or NULL, optionally replace NULL elements with non-NULL character value.

adjustRgb
numeric value adjustment used during the conversion of RGB colors to ANSI colors, which is inherently lossy. If not defined, it uses the default returned by setCLranges() which itself uses getOption("jam.adjustRgb") with default=0. In order to boost color contrast, an alternate value of -0.1 is suggested.

byLine
logical whether to delimit lists by line instead of using collapse to combine them onto one line.

verbose
logical whether to print verbose output

indent
character optional characters used as a prefix to indent output.

file
passed to cat, to allow sending output to a specified file.

append
logical whether to append output, relevant only when file specifies a filename.

invert
logical indicating whether foreground and background colors should be switched.

htmlOut
logical indicating whether to print HTML span output, using format <span style="color:fg;background-color:bg">text</span>.

Value

This function is called for the by-product of printing debug output, it returns invisible(NULL), no output.

Details

This function prints colorized output to the R console, with some rules for colorizing the output to help visually distinguish items. Its output also by default begins with comment '#' characters, a datetimestamp, so it the output is copied back into the R console it will not cause a new command to be run.

The colorization uses a vector or list of colors for fgText (foreground) and bgText (background.), applied to each item in '...'. When an item in '...' is a vector, each vector element is colored alternating light and dark from that base color, to give visual indication of each element. The next item in '...' receives the next color from fgText, and so on. Colors in fgText are recycled to the length of '...'

See also

Examples

printDebug("Testing ", "default ", "printDebug().");

#> ##  (15:25:50) 10Feb2020:  Testing default printDebug(). 

printDebug("List of vectors:", c("one", "two", "three"));

#> ##  (15:25:50) 10Feb2020:  List of vectors:one,two,three 

printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"), collapse=" ");

#> ##  (15:25:50) 10Feb2020:  List of vectors: one, two, three four, five, six 

# slightly different style, one entry per line, indented:
printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"), collapse="\n   ");

#> ##  (15:25:50) 10Feb2020:  List of vectors:
#>    one,
#>    two,
#>    three
#>    four,
#>    five,
#>    six 

# in an R console, or when writing to a log file, the
# following output text is colored
printDebug(c("red", "blue", "yellow"));

#> ##  (15:25:50) 10Feb2020:  red,blue,yellow