rbind a list of vectors into matrix or data.frame

rbind a list of vectors into matrix or data.frame

rbindList(
  x,
  emptyValue = "",
  nullValue = NULL,
  keepListNames = TRUE,
  newColnames = NULL,
  newRownames = NULL,
  fixBlanks = TRUE,
  returnDF = FALSE,
  verbose = FALSE,
  ...
)

Arguments

x
input list of vectors.

emptyValue
character value to use to represent missing values, whenever a blank cell is introduced into the resulting matrix

nullValue
optional value used to replace NULL entries in the input list, useful especially when the data was produced by strsplit() with "". Use nullValue="" to replace NULL with "" and preserve the original list length. Otherwise when nullValue=NULL any empty entries will be silently dropped.

keepListNames
logical whether to use list names as rownames in the resulting matrix or data.frame.

newColnames
NULL or character vector of colnames to use for the resulting matrix or data.frame.

newRownames
NULL or character vector of rownames to use for the resulting matrix or data.frame. If supplied, this value overrides the keepListNames=TRUE use of list names as rownames.

fixBlanks
logical whether to use blank values instead of repeating each vector to the length of the maximum vector length when filling each row of the matrix or data.frame.

returnDF
logical whether to return a data.frame, by default FALSE, a matrix is returned.

verbose
logical whether to print verbose output during processing.

Value

By default a matrix, or if returnDF=TRUE the object is coerced to a data.frame. The rownames by default are derived from the list names, but the colnames are not derived from the vector names.

Details

The purpose of this function is to emulate do.call(rbind, x) on a list of vectors, while specifically handling when there are different numbers of entries per vector. Instead of repeating the values to fill the number of resulting columns, this function fills cells with blank fields. In extensive timings tests at the time this function was created, this method was notably faster than alternatives. That is, it runs do.call(rbind, x) then subsequent steps to replace values with blank entries was notably faster than other alternatives.

See also

Examples

L <- list(a=LETTERS[1:4], b=letters[1:3]);
do.call(rbind, L);

#> Warning: number of columns of result is not a multiple of vector length (arg 2)

#>   [,1] [,2] [,3] [,4]
#> a "A"  "B"  "C"  "D" 
#> b "a"  "b"  "c"  "a" 

rbindList(L);

#>   [,1] [,2] [,3] [,4]
#> a "A"  "B"  "C"  "D" 
#> b "a"  "b"  "c"  ""  

rbindList(L, returnDF=TRUE);

#> Error in unlistDataFrame(as.data.frame(xDF), verbose = verbose, ...): could not find function "unlistDataFrame"