Reload Rmarkdown cache

Reload Rmarkdown cache in the order files were created

reload_rmarkdown_cache(
  dir = ".",
  maxnum = 1000,
  max_cache_name = NULL,
  envir = globalenv(),
  file_sort = c("ctime", "mtime"),
  dryrun = FALSE,
  verbose = TRUE,
  ...
)

Arguments

dir

character path to the directory that contains Rmarkdown cache files. Each file is recognized by the file extension ".rdx".

maxnum

integer indicating the maximum number of cache files to re-load, in order.

max_cache_name

character optional string indicating the name of an Rmarkdown cache chunk where this function will stop loading cache data. All cache files after this point will not be loaded. This option is intended to help recreate the data available to a particular Rmarkdown chunk during processing.

envir

environment where cache data will be loaded.

file_sort

character string indicating how to sort cache files to place them in proper order for re-loading. Note that file modification time mtime may incorrectly sort files that were modified after the initial processing of the Rmarkdown file.

ctime sorts by file creation time, default

mtime sorts by file modification time

dryrun

logical indicating whether to perform a dry-run, which prints messages but does not process the data.

verbose

logical indicating whether to print verbose output.

...

additional arguments are passed to lazyLoad().

Value

this function does not return data, but instead is called for the by-product of loading data into the given envirenvironment.

Details

This function is intended to help re-load Rmarkdown cache files created during the processing/rendering of an Rmarkdown file.

By default, all cached R objects are loaded into the global environment globalenv().

It can be given an argumen envir to store R objects inside a specific environment.

If supplied with max_cache_name then this function will only load the cache chunks in order, until it recognizes that chunk name. This option is intended to help restore the R data available for a particular Rmarkdown chunk.