Rename columns in a data.frame or matrix

Rename columns in a data.frame or matrix

renameColumn(x, from, to, verbose = FALSE, ...)

Arguments

x
data.frame or matrix

from
character vector of colnames expected to be in x. Any values that do not match colnames(x) are ignored.

to
character vector with length(to) == length(from) corresponding to the target name for any colnames that match from.

Value

data.frame (or tibble or DataFrame) with columns from renamed to to as matched. When colnames(x) do not match values in from the corresponding values in from are ignored.

Details

This function is intended to rename one or more columns in a data.frame or matrix. It will gracefully ignore columns which do not match, in order to make it possible to call the function multiple times without problem. It makes it possible to verify that columns are properly named by re-running the function as needed.

This function will also recognize input objects GRanges, ucscData, and IRanges, which store annotation in DataFrame accessible via values(x).

See also

Examples

df <- data.frame(A=1:5, B=6:10, C=11:15);
df;

#>   A  B  C
#> 1 1  6 11
#> 2 2  7 12
#> 3 3  8 13
#> 4 4  9 14
#> 5 5 10 15

df2 <- renameColumn(df,
   from=c("A","C"),
   to=c("a_new", "c_new"));
df2;

#>   a_new  B c_new
#> 1     1  6    11
#> 2     2  7    12
#> 3     3  8    13
#> 4     4  9    14
#> 5     5 10    15

df3 <- renameColumn(df2,
   from=c("A","C","B"),
   to=c("a_new", "c_new","b_new"));
df3;

#>   a_new b_new c_new
#> 1     1     6    11
#> 2     2     7    12
#> 3     3     8    13
#> 4     4     9    14
#> 5     5    10    15