Remove outlier points per row by MAD factor threshold

Remove outlier points per row by MAD factor threshold

rowRmMadOutliers(
  x,
  madFactor = 5,
  na.rm = TRUE,
  minDiff = 0,
  includeAttributes = FALSE,
  verbose = FALSE,
  ...
)

Arguments

x
numeric matrix

madFactor
numeric value to multiply by each row MAD to define the threshold for outlier.

na.rm
logical indicating whether to ignore NA values when calculating the MAD value. It should probably always be TRUE, however setting to FALSE will prevent any calculations in rows that contain NA values, which could be useful.

minDiff
numeric value indicating the minimum difference from median to qualify as an outlier. This value protects against removing outliers which are already extremely close together.

includeAttributes
logical indicating whether to return attributes that describe the threshold and type of threshold used per row, in addition to the madFactor and minDiff values defined.

verbose
logical indicating whether to print verbose output.

...
additional parameters are ignored.

Value

A numeric matrix is returned, with the same dimensions as the input x matrix. Outliers are replaced with NA.

If includeAttributes=TRUE then four attributes will be included as well:

attr(x, "rowThresholds")

The row difference from median threshold used for filtering.

attr(out, "rowThresholds")

When the difference from median is above minDiff it returns "madFactor", otherwise it returns "minDiff".

attr(x, "madFactor")

The madFactor value used.

attr(x, "minDiff")

The minDiff value used.

Details

This function applies outlier detection and removal per row of the input numeric matrix. It first calculates MAD per row, then defines a cutoff threshold by multiplying the MAD by the supplied madFactor, and by ensuring the threshold is at least as large as minDiff.

The minDiff parameter affects cases such as 3 replicates, where all replicates are well within a known threshold indicating low variance, but where two replicates might be nearly identical. Consider c(2.0, 2.0, 2.001). The MAD is zero, therefore any difference from median might otherwise be considered an outlier. One alternative is to define minDiff <- median(rowMads(x)) which sets the minimum threshold to the median MAD across all rows. Such a threshold will only be reasonable if the variance across all rows is expected to be similar.

Note that this function is substantially faster when the matrixStats package is installed, but will use the apply(x, 1, mad) format as a last option.

This function assumes the input data is appropriate for the use of MAD as a summary statistic, specifically that the numeric values per row are expected to be roughly normally distributed, with the occassional presence of outlier points. The outlier points are assumed to be technical outliers, which are often caused by some instrument measurement failure or other upstream protocol failure. A threshold of 5xMAD is fairly lenient and yet is usually sufficient to identify the most egregious outliers. Among technical replicates, often 2xMAD is more appropriate.

See also

Examples

set.seed(123);
x <- matrix(ncol=5, rnorm(25))*5 + 10;
## Define some outlier points
x[1:2,3] <- x[1:2,3]*5 + 50;
x[2:3,2] <- x[2:3,2]*5 - 100;
rowRmMadOutliers(x, madFactor=5);

#>           [,1]      [,2]      [,3]       [,4]     [,5]
#> [1,]  7.197622 18.575325        NA 18.9345657 4.660881
#> [2,]  8.849113        NA        NA 12.4892524 8.910125
#> [3,] 17.793542        NA 12.003857  0.1669142 4.869978
#> [4,] 10.352542  6.565736 10.553414 13.5067795 6.355544
#> [5,] 10.646439  7.771690  7.220794  7.6360430 6.874804