Get chroma (C) and luminance (L) ranges for the given lightMode

Return Crange, Lrange, and adjustRgb values for the given lightMode, which helps define sensible default ranges for contrasting colors.

setCLranges(
  lightMode = checkLightMode(),
  Crange = getOption("jam.Crange", c(30, 190)),
  Lrange = getOption("jam.Lrange"),
  Cgrey = getOption("jam.Cgrey", 5),
  adjustRgb = getOption("jam.adjustRgb", 0),
  setOptions = c("ifnull", "TRUE", "FALSE"),
  verbose = FALSE,
  ...
)

Arguments

lightMode
boolean indicating whether the background color is light (TRUE is bright), or dark (FALSE is dark.) By default it calls checkLightMode() which queries getOption("lightMode").

Crange
numeric range of chroma values, ranging between 0 and 100. By default, getOptions("Crange") is used, otherwise defaults will be assigned based upon lightMode.

Lrange
numeric range of luminance values, ranging between 0 and 100. By default, getOptions("Crange") is used, otherwise defaults will be assigned based upon lightMode.

Cgrey
numeric chroma (C) value, which defines grey colors at or below this chroma. Any colors at or below the grey cutoff will have their C values unchanged. This mechanism prevents converting black to red, for example. To disable the effect, set Cgrey=-1.

adjustRgb
numeric color adjustment factor, used during the conversion of RGB colors to the ANSI-compatible colors used by the crayon pacakge. The ANSI color range does not include a full RGB palette, and the conversion is somewhat lossy. By default, getOptions("jam.adjustRgb") is used to store a globally re-usable value.

setOptions
character or logical whether to update Crange and Lrange options during the subsequent call to setCLranges(). By default, "ifnull" will update only options which were previously NULL; "FALSE" prevents modifying the global options; "TRUE" will update these options with the current values.

verbose
boolean indicating whether to print verbose output.

...
additional parameters are ignored.

Value

list with elements

Crange

Numeric vector of length 2, defining the chroma (C) range.

Lrange

Numeric vector of length 2, defining the luminance (L) range.

adjustRgb

Numeric vector of length 1, defining the adjustment to apply during RGB-to-ANSI color conversion.

Cgrey

Numeric vector of length 1, defining the chroma value below which colors are considered greyscale, where chroma ranges from 0 to 100. Value of -1 (or FALSE) disables this logic.

Details

This function is intended mainly for internal use by jamba such as printDebug(), and make_styles(). It is split into its own function in order to help share the logic of determining some sensible default ranges.

See also

Examples

setCLranges(lightMode=FALSE)

#> $Crange
#> [1]  30 190
#> 
#> $Lrange
#> [1]  5 70
#> 
#> $adjustRgb
#> [1] 0
#> 
#> $Cgrey
#> [1] 5
#>